refactoring clients
