+++ cg_lims/clients/cg/status_db_api.py
@@ -3,8 +3,7 @@ import logging
 from typing import Any, Dict, List
 
 import requests
-from requests import Response
-
+from cg_lims.clients.cg.token_manager import TokenManager
 from cg_lims.exceptions import (
     CgAPIClientConnectionError,
     CgAPIClientDecodeError,
@@ -13,7 +12,7 @@ from cg_lims.exceptions import (
     LimsError,
 )
 from cg_lims.models.sample_lane_sequencing_metrics import SampleLaneSequencingMetrics
-from cg_lims.token_manager import TokenManager
+from requests import Response
 
 LOG = logging.getLogger(__name__)
