+++ cg_lims/commands/base.py
@@ -4,12 +4,12 @@ from typing import Dict
 import click
 import yaml
 from cg_lims import options
+from cg_lims.clients.cg.status_db_api import StatusDBAPI
+from cg_lims.clients.cg.token_manager import TokenManager
 
 # commands
 from cg_lims.EPPs import epps
 from cg_lims.scripts.base import scripts
-from cg_lims.status_db_api import StatusDBAPI
-from cg_lims.token_manager import TokenManager
 from genologics.lims import Lims
